Remove asset size metrics from deploy-hashed-assets / deploy-static
This is tricky to do now that FT Graphite v2 requires every client
that is sending metrics to have an app UUID. Given that `nht` is
a CLI tool that has been sending metrics from CI environments,
this starts to get a bit messy.

We've decided to drop the sending of metrics from `nht` as there
is ongoing work to improve how we deploy and how we build the UI
for applications. Hopefully those efforts will be able to
reintroduce asset size metrics in some other form.