Override default maximum size of HTTP headers
Node.js 8.14.0 introduced a maximum size for HTTP headers of 8KB.
This ended up breaking a lot of our apps because our header sizes are
much larger. In Node.js 8.15.0, you can specify a different maximum size
using the `--max-http-header-size` option.

While adding a smoke test to next-router to test the maximum header
size, we found that the test wasn't passing locally because of
`nht run`. Turns out we need to add the
`node --max-http-header-size=80000` option when starting node and that's
what this PR does!

 🐿 v2.12.0