Add upload-assets-to-s3 task
This adds a new tool which can find files within a directory via a glob match and upload them to S3 with the correct file type and context encoding.

It is intended to supersede the existing deploy-hashed-assets task which was designed for use with apps built with n-ui.

Example usage:

```bash
nht upload-assets-to-s3 --directory=public --bucket=my-bucket --destination=hashed-assets
```

 🐿 v2.12.4