Upgrade node engine from 8.x to 10.x
Many Next apps are using this on Node 12 so I'm confident it works regardless,
but I'd like to install it in ft-app which enforces engine-strict = true and
runs Node 10.

Unit tests pass on Node 10 and will run on CircleCI.

I just need one release that supports Node 10 and then I'm happy for
someone to let it run on Node 12 if they'd like to.

Node 8 is end-of-life as of 1 Jan 2020 so I think we're happy to drop
support for it.